requirements.txt
Django>=4.2
django-unfold